Update query logic
