+++ app/services/vacancy_counter.rb
@@ -14,15 +14,24 @@ class VacancyCounter
 
   class << self
     def role_counts(scope:)
-      scope.map(&:job_roles).flatten.group_by(&:itself).symbolize_keys.transform_values(&:count)
+      scope.joins("CROSS JOIN LATERAL unnest(job_roles) AS r(role)")
+           .group("r.role")
+           .count
+           .transform_keys { |k| Vacancy::JOB_ROLES.invert[k].to_sym }
     end
 
     def phase_counts(scope:)
-      scope.map(&:phases).flatten.group_by(&:itself).symbolize_keys.transform_values(&:count)
+      scope.joins("CROSS JOIN LATERAL unnest(phases) AS p(phase)")
+           .group("p.phase")
+           .count
+           .transform_keys { |k| Vacancy::PHASES.invert.fetch(k) }
     end
 
     def working_pattern_counts(scope:)
-      scope.map(&:working_patterns).flatten.group_by(&:itself).symbolize_keys.transform_values(&:count)
+      scope.joins("CROSS JOIN LATERAL unnest(working_patterns) AS w(pattern)")
+           .group("w.pattern")
+           .count
+           .transform_keys { |k| Vacancy::WORKING_PATTERNS_ENUM.invert[k] }
     end
 
     def job_share_counts(scope:)
@@ -30,7 +39,10 @@ class VacancyCounter
     end
 
     def subject_counts(scope:)
-      scope.map(&:subjects).flatten.compact.group_by(&:itself).symbolize_keys.transform_values(&:count).tap do |hash|
+      scope.joins("CROSS JOIN LATERAL unnest(subjects) AS s(subject)")
+           .group("s.subject")
+           .count
+           .symbolize_keys.tap do |hash|
         GROUPED_SUBJECTS.each do |group_name, subjects|
           hash[group_name] = hash.slice(*subjects + [group_name]).values.sum
         end
